// Practices/Train_Heart_Failure/app_front/heart-failure-front/src/app/common/test-data.interface.ts
export interface TestData {
    Age: number;
    Sex: string;
    ChestPainType: string;
    RestingBP: number;
    Cholesterol: number;
    FastingBS: number;
    RestingECG: string;
    MaxHR: number;
    ExerciseAngina: string;
    Oldpeak: number;
    ST_Slope: string;
    HeartDisease: number;
  }
  